refactor: should-skip-query moved to workspace utils
