+++ packages/twenty-server/src/workspace/workspace-resolver-builder/workspace-resolver.factory.ts
@@ -8,7 +8,7 @@ import { getResolverName } from 'src/workspace/utils/get-resolver-name.util';
 import { UpdateManyResolverFactory } from 'src/workspace/workspace-resolver-builder/factories/update-many-resolver.factory';
 import { DeleteManyResolverFactory } from 'src/workspace/workspace-resolver-builder/factories/delete-many-resolver.factory';
 import { ExecuteQuickActionOnOneResolverFactory } from 'src/workspace/workspace-resolver-builder/factories/execute-quick-action-on-one-resolver.factory';
-import { shouldSkipQuery } from 'src/workspace/workspace-schema-builder/utils/should-skip-query.util';
+import { shouldSkipQuery } from 'src/workspace/utils/should-skip-query.util';
 
 import { FindDuplicatesResolverFactory } from './factories/find-duplicates-resolver.factory';
 import { FindManyResolverFactory } from './factories/find-many-resolver.factory';
